Create a simple travel diary for a Monash trip

Creates a simple travel diary (consisting of one destination) for a trip. Used in conjunction with render_travel_diary() to generate a PDF travel diary for uploading.

create_simple_travel_diary(
  start_date,
  end_date,
  start_city = "Melbourne",
  destination,
  reason
)

Arguments

start_date

The start date of the trip (as a Date object).

end_date

The end date of the trip (as a Date object).

start_city

The city where the trip starts (default is "Melbourne").

destination

The destination city for the trip.

reason

The reason for the trip (e.g., "Conference", "Meeting").

Value

A data frame representing the travel diary, with columns for Date, Location, and Description.

Examples

create_simple_travel_diary(
  start_date = as.Date("2025-08-18"), 
  end_date = as.Date("2025-09-01"), 
  destination = "Perth", 
  reason = "NUMBATS Conference")
#>          Date  Location         Description
#> 1  2025-08-18 Melbourne     Travel to Perth
#> 2  2025-08-19     Perth  NUMBATS Conference
#> 3  2025-08-20     Perth  NUMBATS Conference
#> 4  2025-08-21     Perth  NUMBATS Conference
#> 5  2025-08-22     Perth  NUMBATS Conference
#> 6  2025-08-23     Perth  NUMBATS Conference
#> 7  2025-08-24     Perth  NUMBATS Conference
#> 8  2025-08-25     Perth  NUMBATS Conference
#> 9  2025-08-26     Perth  NUMBATS Conference
#> 10 2025-08-27     Perth  NUMBATS Conference
#> 11 2025-08-28     Perth  NUMBATS Conference
#> 12 2025-08-29     Perth  NUMBATS Conference
#> 13 2025-08-30     Perth  NUMBATS Conference
#> 14 2025-08-31     Perth  NUMBATS Conference
#> 15 2025-09-01     Perth Travel to Melbourne