likert {monash}	R Documentation

Likert-scale data

Description

Create Likert-scale data object suitable for plotting using plot.likert and calculating descriptive measures and quantiles from grouped data

Usage

likert(x, upper = max(x), labels)

Arguments

x 	 The observed response (number of observations in each category of likert-scale response)

upper 	 Number of permissible categories (number of categories for each likert-scale question)

labels 	 Labels of likert-scale respose. By default, labels of 5-point likert-scale question are SD=Strongly Disagree, D=Disagree, N=Neutral, A=Agree, SA=Strongly Agree

Details

Because Likert-scale data consist of discrete values, quantiles calculated in the usual way can be misleading. One solution to this problem is to compute pseudo-quantiles based on the distribution function. We assume that there is an underlying continuous variable which is observed to the nearest integer. Then the mid point quantile estimator from grouped data (e.g., Schmeiser and Deutsch, 1977) can be used.

Value

An object of class "likert" is a list containing the following elements

Number 	 Frequency in each category

Percent 	 Percentage in each category

data 	 Observed response

upper 	 Maximum number of categories

labels 	 Labels of categories

Author(s)

Rob J Hyndman

References

Schmeiser, B.W., and Deutsch, S.J. (1977). Quantile estimation from grouped data: the cell mid point. Comm. Stat.-sim.comp., B6(3), 221-234.

See Also

See Also mean.likert, median.likert, plot.likert, quantile.likert, sd.likert, summary.likert, table.likert

Examples

## Let x be a set of responses to 5-point Likert scale questions
x <- c(rep(3,9),rep(4,7),rep(5,14))
y <- likert(x)
quantile(y)
plot(y)

[Package monash version 1.02 Index]