ls.size {monash}	R Documentation

List objects in size order

Description

Function to list all objects in the current environment in order of size from largest to smallest.

Usage

ls.size()

Author(s)

Rob J Hyndman

See Also

ls,object.size

[Package monash version 1.02 Index]