median {monash}	R Documentation

Median

Description

Generic function for the median.

Usage

median(...)

Arguments

... 	 Arguments passed to specific methods.

Details

The median function in the stats package is replaced by median.default.

Value

Refer to specific methods. For numeric vectors, see the median function in the stats package.

Author(s)

Rob J Hyndman

See Also

median.likert

[Package monash version 1.02 Index]