monash {monash}	R Documentation

Functions for use by Monash University Statistics and Econometrics Consulting Services

Description

Functions for saving EPS figures for use in Latex, for summarizing and plotting Likert-scale data, etc.

Details

Package: 	 monash

Type: 	 Package

Version: 	 1.0

Date: 	 2005-11-28

License: 	 GPL version 2 or newer

Author(s)

Rob J. Hyndman
Maintainer: Muhammad Akram
<Muhammad.Akram@buseco.monash.edu.au>

[Package monash version 1.02 Index]