sd {monash}	R Documentation

Standard deviation and variance

Description

Generic functions for standard deviation and variance.

Usage

sd(...)
var(...)

Arguments

... 	 Arguments passed to specific methods.

Details

The sd and var functions in the stats package are replaced by sd.default and var.default.

Value

Refer to specific methods. For numeric vectors, see the sd and var functions in the stats package.

Author(s)

Rob J Hyndman

See Also

sd.likert, var.likert

[Package monash version 1.02 Index]