savefig {monash}	R Documentation

Save figure to file

Description

savefig saves figures to files with minimal surrounding white space, suitable for inclusion in books and reports. savefig is especially good for including plots in LaTeX. File formats supported are eps, pdf and jpg. Default file format for savefig is eps. The other functions are wrappers for saving specific file formats.

Usage

savefig(filename, height = 10, width = (1 + sqrt(5))/2*height, 
    type=c("eps","pdf","jpg"), pointsize = 10, family = "Helvetica", 
    sublines = 0, toplines = 0, leftlines = 0) 
saveeps(...)
savepdf(...)
savejpg(...)
savepng(...)

Arguments

filename 	 File in which to store the figure. Adds ".eps" ".pdf" or ".jpg" to the argument

width 	 Width of figure in centimetres. Default 15cm.

height 	 Height of figure in centimetres. Default 1.618*height.

type 	 File format. Either "eps", "pdf" or "jpg". Default is "eps".

pointsize 	 The default point size to be used.

family 	 Font family. See postscript for details.

sublines 	 Number of lines to reserve for subtitles.

toplines 	 Number of lines to reserve for main title.

leftlines 	 Number of lines to reserve for left margin.

... 	 All arguments passed to savefig.

Author(s)

Rob J Hyndman

See Also

postscript, pdf, jpeg

[Package monash version 1.02 Index]