statfun.likert {monash}	R Documentation

Descriptive statistics for Likert-scale data

Description

Functions to calculate descriptive statistics of Likert-scale data including mean, median, quantile, standard deviation, variance and semi-interquartile-range.

Usage

## S3 method for class 'likert':
mean(x, na.rm=FALSE, ...)
## S3 method for class 'likert':
median(x, na.rm=FALSE, ...)
## S3 method for class 'likert':
quantile(x, probs=seq(0, 1, 0.25), truncate=FALSE, ...)
## S3 method for class 'likert':
sd(x, na.rm=FALSE, ...)
## S3 method for class 'likert':
var(x, na.rm=FALSE, ...)
sir(x)

Arguments

x 	 An object of class likert.

probs 	 numeric vector of probabilities with values in [0,1].

truncate 	 A logical value, if TRUE, the quantile can not be outside the range [1,5].

na.rm 	 A logical value indicating whether NA values should be stripped before the computation proceeds.

... 	 Other arguments

Details

Because Likert-scale data consist of discrete values, quantiles calculated in the usual way can be misleading. One solution to this problem is to compute pseudo-quantiles based on the distribution function. We assume that there is an underlying continuous variable which is observed to the nearest integer. Then the mid point quantile estimator from grouped data (e.g., Schmeiser and Deutsch, 1977) can be used.

Value

Corresponding measure of likert-scale data

Author(s)

Rob J Hyndman

References

Schmeiser, B.W., and Deutsch, S.J. (1977). Quantile estimation from grouped data: the cell mid point. Comm. Stat.-sim.comp., B6(3), 221-234.

See Also

See Also likert

Examples

x <- c(rep(3,9),rep(4,7),rep(5,14))
y <- likert(x)
mean(y)
median(y)
quantile(y)
sd(y)
var(y)
sir(y)

[Package monash version 1.02 Index]