summary.likert {monash}	R Documentation

Summary meaasures of Likert scale data

Description

Summarizes Likert-scale data output, including quantiles along with descriptive statistics

Usage

summary.likert(object, ...)

Arguments

object 	 An object of likert-scale data

... 	 Other arguments.

Details

Because Likert-scale data consist of discrete values, quantiles calculated in the usual way can be misleading. One solution to this problem is to compute pseudo-quantiles based on the distribution function. We assume that there is an underlying continuous variable which is observed to the nearest integer. Then the mid point quantile estimator from grouped data (e.g., Schmeiser and Deutsch, 1977) can be used.

Value

An object of class "Likert" is a list containing the following elements

Number 	 Frequency to each category

Percent 	 Percentage of each category

data 	 Observed response

upper 	 Maximum number of categories

labels 	 Labels of categories

Author(s)

Rob J Hyndman

References

Schmeiser, B.W., and Deutsch, S.J. (1977). Quantile estimation from grouped data: the cell mid point. Comm. Stat.-sim.comp., B6(3), 221-234.

See Also

See Also mean.likert, median.likert, plot.likert, quantile.likert, sd.likert, summary.likert, table.likert

Examples

x <- c(rep(3,9),rep(4,7),rep(5,14))
y <- likert(x)
summary(y)

[Package monash version 1.02 Index]