table {monash}	R Documentation

Cross tabulation and table creation

Description

Generic function for creating tables.

Usage

table(...)

Arguments

... 	 Arguments passed to specific methods.

Details

The table function in the base package is replaced by table.default.

Value

Refer to specific methods. For factors, see the table function in the base package.

Author(s)

Rob J Hyndman

See Also

table.likert

[Package monash version 1.02 Index]